+++ client/src/app/app.module.ts
@@ -16,9 +16,9 @@ import {BlindsCommandService} from './remote/blinds-command.service';
 import {BlindsComponent} from './blinds/blinds.component';
 import {TemperatureComponent} from './temperature/temperature.component';
 import {TemperatureViewComponent} from './temperature-view/temperature-view.component';
-import {BlindsDataObservablePipe} from './blinds/blinds-data-observable.pipe';
-import {BlindsDataFormatterPipe} from './blinds/blinds-data-formatter.pipe';
-import {BlindsPercentageDownPipe} from './blinds/blinds-percentage-down.pipe';
+import {BlindsDataObservablePipe} from './blinds/pipes/blinds-data-observable.pipe';
+import {BlindsDataFormatterPipe} from './blinds/pipes/blinds-data-formatter.pipe';
+import {BlindsPercentageDownPipe} from './blinds/pipes/blinds-percentage-down.pipe';
 import {InfoComponent} from './info/info.component';
 import {CommonRestService} from './remote/common-rest.service';
 import {BlindsButtonsComponent} from './blinds/blinds-buttons/blinds-buttons.component';
