+++ client/src/app/blinds/pipes/blinds-data-formatter.pipe.spec.ts
