+++ client/src/app/blinds/pipes/blinds-data-observable.pipe.spec.ts
