+++ client/src/app/blinds/pipes/blinds-percentage-down.pipe.spec.ts
