+++ client/src/app/blinds/pipes/blinds-percentage-down.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import {IBlindsData} from '../../../../server/entities/data.interface';
-import {blindsStateAsString} from '../../../../server/entities/blinds-state';
+import {IBlindsData} from '../../../../../server/entities/data.interface';
+import {blindsStateAsString} from '../../../../../server/entities/blinds-state';
 
 @Pipe({
   name: 'blindsPercentageDown'
